requirements.txt
pandas
numpy
streamlit
matplotlib
playwright
colorlog
openpyxl
dotenv
openai
bs4
requests
readability-lxml
